reorganizes admin,validators,enums as part of events app
