+++ physionet-django/events/enums.py
@@ -1,0 +1,6 @@
+from django.db import models
+
+
+class EventCategory(models.TextChoices):
+    COURSE = "Course"
+    WORKSHOP = "Workshop"
